tests/wpt/tests/html/canvas/element/drawing-images-to-the-canvas/2d.drawImage.5arg.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>Canvas test: 2d.drawImage.5arg</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>
<link rel="stylesheet" href="/html/canvas/resources/canvas-tests.css">
<body class="show_output">

<h1>2d.drawImage.5arg</h1>
<p class="desc"></p>


<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
promise_test(async t => {

  var canvas = document.getElementById('c');
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  const response_red = await fetch('/images/red.png');
  const blob_red = await response_red.blob();
  const bitmap_red = await createImageBitmap(blob_red);

  const response_green = await fetch('/images/green.png');
  const blob_green = await response_green.blob();
  const bitmap_green = await createImageBitmap(blob_green);

  ctx.drawImage(bitmap_green, 50, 0, 50, 50);
  ctx.drawImage(bitmap_red, 0, 0, 50, 50);
  ctx.fillStyle = '#0f0';
  ctx.fillRect(0, 0, 50, 50);
  _assertPixelApprox(canvas, 0,0, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 99,0, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 0,49, 0,255,0,255, 2);
  _assertPixelApprox(canvas, 99,49, 0,255,0,255, 2);

}, "");
</script>
